+++ src/GotimeServiceProvider.php
@@ -132,7 +132,7 @@ class GotimeServiceProvider extends ServiceProvider
 
             // layouts
             $this->registerComponentX('layouts.admin-form', 'admin-form');
-            Blade::component('gotime::layouts.base', 'gotime-layouts.base');
+            Blade::component('gotime::components.layouts.base', 'gotime-layouts.base');
 
             // table components
             $this->registerComponentX('table.th', 'table.th');
@@ -170,6 +170,7 @@ class GotimeServiceProvider extends ServiceProvider
         $this->registerComponentX('input.textarea', 'textarea');
 
         // Combined Controls
+        $this->registerComponentX('input-v2.radio', 'radio');
         $this->registerComponentX('input.checkbox', 'checkbox');
         $this->registerComponentX('input.email');
         $this->registerComponentX('input.file');
@@ -186,7 +187,6 @@ class GotimeServiceProvider extends ServiceProvider
         $this->registerComponentX('input.filepond', 'filepond');
         $this->registerComponentX('input.trix', 'trix');
     }
-
 
 
     /**